Moved DownloadFile from 'resources' package into 'httpclient'
