+++ arduino/resources/download.go
@@ -18,9 +18,7 @@ package resources
 import (
 	"fmt"
 	"os"
-	"time"
 
-	"github.com/arduino/arduino-cli/arduino"
 	"github.com/arduino/arduino-cli/arduino/httpclient"
 	paths "github.com/arduino/go-paths-helper"
 	"go.bug.st/downloader/v2"
@@ -28,7 +26,7 @@ import (
 
 // Download performs a download loop using the provided downloader.Downloader.
 // Messages are passed back to the DownloadProgressCB using label as text for the File field.
-func (r *DownloadResource) Download(downloadDir *paths.Path, config *downloader.Config, label string, downloadCB DownloadProgressCB) error {
+func (r *DownloadResource) Download(downloadDir *paths.Path, config *downloader.Config, label string, downloadCB httpclient.DownloadProgressCB) error {
 	path, err := r.ArchivePath(downloadDir)
 	if err != nil {
 		return fmt.Errorf(tr("getting archive path: %s"), err)
@@ -47,7 +45,7 @@ func (r *DownloadResource) Download(downloadDir *paths.Path, config *downloader.
 			// File is cached, nothing to do here
 
 			// This signal means that the file is already downloaded
-			downloadCB(&DownloadProgress{
+			downloadCB(&httpclient.DownloadProgress{
 				File:      label,
 				Completed: true,
 			})
@@ -56,71 +54,5 @@ func (r *DownloadResource) Download(downloadDir *paths.Path, config *downloader.
 	} else {
 		return fmt.Errorf(tr("getting archive file info: %s"), err)
 	}
-	return DownloadFile(path, r.URL, label, downloadCB, config)
+	return httpclient.DownloadFile(path, r.URL, label, downloadCB, config)
 }
-
-// DownloadFile downloads a file from a URL into the specified path. An optional config and options may be passed (or nil to use the defaults).
-// A DownloadProgressCB callback function must be passed to monitor download progress.
-func DownloadFile(path *paths.Path, URL string, label string, downloadCB DownloadProgressCB, config *downloader.Config, options ...downloader.DownloadOptions) error {
-	if config == nil {
-		c, err := GetDownloaderConfig()
-		if err != nil {
-			return err
-		}
-		config = c
-	}
-
-	d, err := downloader.DownloadWithConfig(path.String(), URL, *config, options...)
-	if err != nil {
-		return err
-	}
-	downloadCB(&DownloadProgress{
-		File:      label,
-		URL:       d.URL,
-		TotalSize: d.Size(),
-	})
-
-	err = d.RunAndPoll(func(downloaded int64) {
-		downloadCB(&DownloadProgress{Downloaded: downloaded})
-	}, 250*time.Millisecond)
-	if err != nil {
-		return err
-	}
-
-	// The URL is not reachable for some reason
-	if d.Resp.StatusCode >= 400 && d.Resp.StatusCode <= 599 {
-		return &arduino.FailedDownloadError{Message: tr("Server responded with: %s", d.Resp.Status)}
-	}
-
-	downloadCB(&DownloadProgress{Completed: true})
-	return nil
-}
-
-// GetDownloaderConfig returns the downloader configuration based on current settings.
-func GetDownloaderConfig() (*downloader.Config, error) {
-	httpClient, err := httpclient.New()
-	if err != nil {
-		return nil, &arduino.InvalidArgumentError{Message: tr("Could not connect via HTTP"), Cause: err}
-	}
-	return &downloader.Config{
-		HttpClient: *httpClient,
-	}, nil
-}
-
-// DownloadProgress is a report of the download progress, not all fields may be
-// filled and multiple reports may be sent during a download.
-type DownloadProgress struct {
-	// URL of the download.
-	URL string
-	// The file being downloaded.
-	File string
-	// TotalSize is the total size of the file being downloaded.
-	TotalSize int64
-	// Downloaded is the size of the downloaded portion of the file.
-	Downloaded int64
-	// Completed reports whether the download is complete.
-	Completed bool
-}
-
-// DownloadProgressCB is a callback function to report download progress
-type DownloadProgressCB func(progress *DownloadProgress)
